qv-lagar-cetagua-backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.7
django-cors-headers==4.3.1
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
python-dotenv==1.0.1